Move reset 2sv test into shared managing 2sv test file
