+++ test/integration/managing_two_step_verification_test.rb
@@ -9,11 +9,15 @@ class ManagingTwoStepVerificationTest < ActionDispatch::IntegrationTest
     @child_organisation = create(:organisation, parent: @organisation)
   end
 
-  context "Mandating 2-step verification" do
+  context "Mandating and resetting 2-step verification" do
     setup do
       @user = create(:user, organisation: @organisation)
+      @user_requring_2sv = create(:two_step_enabled_user, organisation: @organisation)
       @user_in_child_organisation = create(:user, organisation: @child_organisation)
+      @user_requring_2sv_in_child_organisation = create(:two_step_enabled_user, organisation: @child_organisation)
+      @user_in_different_organisation = create(:two_step_enabled_user, organisation: create(:organisation))
     end
+
     context "when logged in as a super admin" do
       setup do
         @super_admin = create(:superadmin_user)
@@ -37,9 +41,13 @@ class ManagingTwoStepVerificationTest < ActionDispatch::IntegrationTest
 
         expected_account_logs = [
           "Exemption from 2-step verification removed by #{@super_admin.name}",
-          "2-step verification setup mandated at next login by #{@super_admin.name}"
+          "2-step verification setup mandated at next login by #{@super_admin.name}",
         ]
         assert_user_access_log_contains_messages(user, expected_account_logs)
+      end
+
+      should "reset 2-step verification and notify the chosen user by email for users in any organisation" do
+        assert_2sv_can_be_reset(@super_admin, @user_requring_2sv)
       end
     end
 
@@ -54,6 +62,10 @@ class ManagingTwoStepVerificationTest < ActionDispatch::IntegrationTest
 
       should "be able to unset the requirement for 2fa" do
         assert_admin_can_remove_2sv_requirement_without_notifying_user(@admin, @user)
+      end
+
+      should "reset 2-step verification and notify the chosen user by email for users in any organisation" do
+        assert_2sv_can_be_reset(@admin, @user_requring_2sv)
       end
     end
 
@@ -77,6 +89,18 @@ class ManagingTwoStepVerificationTest < ActionDispatch::IntegrationTest
       should "be able to unset the requirement for 2fa for a user in a child organisation" do
         assert_admin_can_remove_2sv_requirement_without_notifying_user(@super_org_admin, @user_in_child_organisation)
       end
+
+      should "be able to reset 2-step verification and notify the chosen user by email if they belong to the same org as the user" do
+        assert_2sv_can_be_reset(@super_org_admin, @user_requring_2sv)
+      end
+
+      should "be able to reset 2-step verification and notify the chosen user by email if the user is in a child organisation" do
+        assert_2sv_can_be_reset(@super_org_admin, @user_requring_2sv_in_child_organisation)
+      end
+
+      should "not be able to reset 2-step verification and notify the chosen user by email if the user is in a different organisation" do
+        assert_2sv_cannot_be_reset(@super_org_admin, @user_in_different_organisation)
+      end
     end
 
     context "when logged in as an organisation admin" do
@@ -91,6 +115,18 @@ class ManagingTwoStepVerificationTest < ActionDispatch::IntegrationTest
       should "be able to unset the requirement for 2fa" do
         assert_admin_can_remove_2sv_requirement_without_notifying_user(@org_admin, @user)
       end
+
+      should "be able to reset 2-step verification and notify the chosen user by email if they belong to the same org as the user" do
+        assert_2sv_can_be_reset(@org_admin, @user_requring_2sv)
+      end
+
+      should "not be able to reset 2-step verification and notify the chosen user by email if the user is in a child organisation" do
+        assert_2sv_cannot_be_reset(@org_admin, @user_requring_2sv_in_child_organisation)
+      end
+
+      should "not be able to reset 2-step verification and notify the chosen user by email if the user is in a different organisation" do
+        assert_2sv_cannot_be_reset(@org_admin, @user_in_different_organisation)
+      end
     end
 
     context "when logged in as a normal user" do
@@ -99,6 +135,7 @@ class ManagingTwoStepVerificationTest < ActionDispatch::IntegrationTest
         sign_in_as_and_edit_user(non_admin_user, @user)
 
         assert page.has_no_text? "Mandate 2-step verification for this user"
+        assert page.has_no_link? "Reset 2-step verification"
       end
     end
